app/requirements.txt
Flask
asyncio-nats-client
requests
